Refactor: rewrite tests against Docker containers
