+++ src/test/java/com/suse/salt/netapi/client/SaltClientDockerTest.java
@@ -1,0 +1,118 @@
+package com.suse.salt.netapi.client;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+
+import com.suse.salt.netapi.AuthModule;
+import com.suse.salt.netapi.datatypes.Token;
+import com.suse.salt.netapi.exception.SaltException;
+import com.suse.salt.netapi.exception.SaltUserUnauthorizedException;
+
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+import java.net.URI;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
+
+/**
+ * SaltStack API unit tests against a running SaltStack Master with NetAPI
+ * module enabled. In order to pass these tests you MUST have a salt-master
+ * NetAPI enabled listening at: http://SALT_NETAPI_SERVER:SALT_NETAPI_PASSWORD/
+ *
+ * If you are looking for a quick way to bring up a SaltStack Master with NETAPI
+ * enabled locally, take a look at .travis.yml
+ *
+ */
+public class SaltClientDockerTest {
+
+    private static final String SALT_NETAPI_SERVER = "http://localhost";
+    private static final int SALT_NETAPI_PORT = 8000;
+    private static final String SALT_NETAPI_USER = "saltdev";
+    private static final String SALT_NETAPI_PASSWORD = "saltdev";
+    private static final AuthModule SALT_NETAPI_AUTH = AuthModule.PAM;
+
+    private SaltClient client;
+
+    @Rule
+    public ExpectedException exception = ExpectedException.none();
+
+    @Before
+    public void init() throws SaltException {
+        URI uri = URI.create(SALT_NETAPI_SERVER + ":" + SALT_NETAPI_PORT);
+        client = new SaltClient(uri);
+    }
+
+    @Test
+    public void testLoginOk() throws Exception {
+        Token token =
+                client.login(SALT_NETAPI_USER, SALT_NETAPI_PASSWORD, SALT_NETAPI_AUTH);
+        assertNotNull(token);
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime tokenStart = LocalDateTime.ofInstant(token.getStart().toInstant(),
+                ZoneId.systemDefault());
+        LocalDateTime tokenExpiration = LocalDateTime
+                .ofInstant(token.getExpire().toInstant(), ZoneId.systemDefault());
+
+        assertTrue(tokenStart.isBefore(now));
+        assertTrue(tokenExpiration.isAfter(now));
+    }
+
+    @Test(expected = SaltUserUnauthorizedException.class)
+    public void testLoginFailure() throws Exception {
+        client.login("user", "pass", AuthModule.DJANGO);
+    }
+
+    @Test
+    public void testLoginAsyncOk() throws Exception {
+        Future<Token> futureToken =
+                client.loginAsync(SALT_NETAPI_USER, SALT_NETAPI_PASSWORD, SALT_NETAPI_AUTH);
+        Token token = futureToken.get();
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime tokenStart = LocalDateTime.ofInstant(token.getStart().toInstant(),
+                ZoneId.systemDefault());
+        LocalDateTime tokenExpiration = LocalDateTime
+                .ofInstant(token.getExpire().toInstant(), ZoneId.systemDefault());
+
+        assertTrue(tokenStart.isBefore(now));
+        assertTrue(tokenExpiration.isAfter(now));
+    }
+
+    @Test(expected = ExecutionException.class)
+    public void testLoginAsyncFailure() throws Exception {
+        Future<Token> futureToken = client.loginAsync("user", "pass", AuthModule.DJANGO);
+        Token token = futureToken.get();
+        assertNull(token);
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    public void testGetMinions() throws Exception {
+        client.login(SALT_NETAPI_USER, SALT_NETAPI_PASSWORD, SALT_NETAPI_AUTH);
+        Map<String, Map<String, Object>> minions = client.getMinions();
+
+        assertNotNull(minions);
+        assertEquals(2, minions.size());
+
+        minions.keySet().forEach(minionHostname -> {
+            Map<String, Object> minionInfo = minions.get(minionHostname);
+            assertEquals("Linux", minionInfo.get("kernel"));
+            assertEquals("openSUSE Leap", minionInfo.get("lsb_distrib_id"));
+            assertEquals("42.2", minionInfo.get("lsb_distrib_release"));
+            assertTrue(minionInfo.get("saltversioninfo") instanceof List);
+            List<String> saltVersionInfo = (List<String>) minionInfo.get("saltversioninfo");
+            assertEquals(2016.0, saltVersionInfo.get(0));
+            assertEquals(3.0, saltVersionInfo.get(1));
+            assertEquals(4.0, saltVersionInfo.get(2));
+            assertEquals(0.0, saltVersionInfo.get(3));
+        });
+    }
+}
